requirements.txt
fastapi
pydantic
soda-core
soda-core-contracts
soda-core-spark-df
uvicorn
pandas
pyspark
